@@ -1,116 +1,79 @@
 package com.example.cameraandgallery
 
 import android.Manifest
-import android.app.Activity
 import android.content.ContentValues
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.graphics.Bitmap
+import android.graphics.ImageDecoder
 import android.net.Uri
 import android.os.Build
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.provider.MediaStore
-import android.util.Log
 import android.widget.Toast
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import kotlinx.android.synthetic.main.activity_main.*
-import java.io.*
+import com.example.cameraandgallery.databinding.ActivityMainBinding
+import java.io.IOException
 import java.text.SimpleDateFormat
 
-class MainActivity : AppCompatActivity() {
 
-    val CAMERA_PERMISSION = arrayOf(Manifest.permission.CAMERA)
-    val STORAGE_PERMISSION = arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE
-        , Manifest.permission.WRITE_EXTERNAL_STORAGE)
+class MainActivity : BaseActivity() {
 
-    val FLAG_PERM_CAMERA = 98
-    val FLAG_PERM_STORAGE = 99
+    val PERM_STOREGE = 99
+    val PERM_CAMERA = 100
+    val REQ_CAMERA = 101
+    val REQ_STORAGE = 102
 
-    val FLAG_REQ_CAMERA = 101
-    val FLAG_REQ_STORAGE = 102
+    val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
+
+    var realUri: Uri? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(binding.root)
 
-        if (checkPermission(STORAGE_PERMISSION, FLAG_PERM_STORAGE)) {
-            setViews()
+        requirePermissions(arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE), PERM_STOREGE)
+    }
+
+    override fun permissionGranted(requestCode: Int) {
+        when(requestCode) {
+            PERM_STOREGE -> setViews()
+            PERM_CAMERA -> openCamera()
         }
     }
 
-    fun setViews(){
-        buttonCamera.setOnClickListener {
-            openCamera()
+    override fun permissionDenied(requestCode: Int) {
+        when(requestCode) {
+            PERM_STOREGE -> {
+                Toast.makeText(baseContext, "외부저장소 권한을 승인해야 앱을 사용할 수 있습니다.", Toast.LENGTH_LONG).show()
+                finish()
+            }
+            PERM_CAMERA -> Toast.makeText(baseContext, "카메라 권한을 승인해야 카메라를 사용할 수 있습니다.", Toast.LENGTH_LONG).show()
         }
-        buttonGallery.setOnClickListener {
+    }
+
+    fun setViews() {
+        binding.buttonCamera.setOnClickListener {
+            requirePermissions(arrayOf(Manifest.permission.CAMERA), PERM_CAMERA)
+        }
+        binding.buttonGallery.setOnClickListener {
             openGallery()
         }
     }
 
     fun openCamera() {
-        if (checkPermission(CAMERA_PERMISSION, FLAG_PERM_CAMERA)) {
-            val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
-            startActivityForResult(intent, FLAG_REQ_CAMERA)
+        val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+
+        createImageUri(newFileName(), "image/jpg")?.let { uri ->
+            realUri = uri
+            intent.putExtra(MediaStore.EXTRA_OUTPUT, realUri)
+            startActivityForResult(intent, REQ_CAMERA)
         }
     }
 
     fun openGallery() {
         val intent = Intent(Intent.ACTION_PICK)
         intent.type = MediaStore.Images.Media.CONTENT_TYPE
-        startActivityForResult(intent, FLAG_REQ_STORAGE)
-    }
+        startActivityForResult(intent, REQ_STORAGE)
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if(resultCode == Activity.RESULT_OK) {
-            when(requestCode){
-                FLAG_REQ_CAMERA -> {
-                    if (data?.extras?.get("data") != null) {
-                        val bitmap = data?.extras?.get("data") as Bitmap
-//                        imagePreview.setImageBitmap(bitmap)
-                        val uri = saveImageFile(newFileName(), "image/jpg", bitmap)
-                        imagePreview.setImageURI(uri)
-                    }
-                }
-                FLAG_REQ_STORAGE -> {
-                    val uri = data?.data
-                    imagePreview.setImageURI(uri)
-                }
-            }
-        }
-    }
-
-    fun saveImageFile(filename:String, mimeType:String, bitmap:Bitmap) : Uri? {
-        var values = ContentValues()
-        values.put(MediaStore.Images.Media.DISPLAY_NAME, filename)
-        values.put(MediaStore.Images.Media.MIME_TYPE, mimeType)
-
-        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            values.put(MediaStore.Images.Media.IS_PENDING, 1)
-        }
-
-        val uri = contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values)
-
-        try {
-            if(uri != null) {
-                var descriptor = contentResolver.openFileDescriptor(uri, "w")
-                if(descriptor != null) {
-                    val fos = FileOutputStream(descriptor.fileDescriptor)
-                    bitmap.compress(Bitmap.CompressFormat.JPEG, 100, fos)
-                    fos.close()
-                    if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                        values.clear()
-                        values.put(MediaStore.Images.Media.IS_PENDING, 0)
-                        contentResolver.update(uri, values, null, null)
-                    }
-                }
-            }
-        }catch (e:java.lang.Exception) {
-            Log.e("File", "error=${e.localizedMessage}")
-        }
-        return uri
     }
 
     fun newFileName() : String {
@@ -120,43 +83,46 @@ class MainActivity : AppCompatActivity() {
         return "$filename.jpg"
     }
 
-    /*
-    * 여기서 부터 권한처리 관련 함수
-    */
-    fun checkPermission(permissions: Array<out String>, flag:Int) : Boolean {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            for (permission in permissions) {
-                if (ContextCompat.checkSelfPermission(this, permission) != PackageManager.PERMISSION_GRANTED) {
-                    ActivityCompat.requestPermissions(this, permissions, flag)
-                    return false
-                }
-            }
-        }
-        return true
+    fun createImageUri(filename: String, mimeType: String) : Uri? {
+        var values = ContentValues()
+        values.put(MediaStore.Images.Media.DISPLAY_NAME, filename)
+        values.put(MediaStore.Images.Media.MIME_TYPE, mimeType)
+        return contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values)
     }
 
-    override fun onRequestPermissionsResult(requestCode: Int
-                                            , permissions: Array<out String>
-                                            , grantResults: IntArray) {
-        when (requestCode) {
-            FLAG_PERM_STORAGE -> {
-                for (grant in grantResults) {
-                    if (grant != PackageManager.PERMISSION_GRANTED) {
-                        Toast.makeText(this, "저장소 권한을 승인해야지만 앱을 사용할 수 있습니다.", Toast.LENGTH_LONG).show()
-                        finish()
-                        return
-                    }
-                }
-                setViews()
+    fun loadBitmap(photoUri: Uri): Bitmap? {
+        var image: Bitmap? = null
+        try {
+            image = if (Build.VERSION.SDK_INT > 27) { // Api 버전별 이미지 처리
+                val source: ImageDecoder.Source =
+                        ImageDecoder.createSource(this.contentResolver, photoUri)
+                ImageDecoder.decodeBitmap(source)
+            } else {
+                MediaStore.Images.Media.getBitmap(this.contentResolver, photoUri)
             }
-            FLAG_PERM_CAMERA -> {
-                for (grant in grantResults) {
-                    if (grant != PackageManager.PERMISSION_GRANTED) {
-                        Toast.makeText(this, "카메라 권한을 승인해야지만 카메라를 사용할 수 있습니다.", Toast.LENGTH_LONG).show()
-                        return
+        } catch (e: IOException) {
+            e.printStackTrace()
+        }
+        return image
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (resultCode == RESULT_OK) {
+            when (requestCode){
+                REQ_CAMERA -> {
+                    realUri?.let { uri ->
+                        val bitmap = loadBitmap(uri)
+                        binding.imagePreview.setImageBitmap(bitmap)
+
+                        realUri = null
                     }
                 }
-                openCamera()
+                REQ_STORAGE -> {
+                    data?.data?.let { uri ->
+                        binding.imagePreview.setImageURI(uri)
+                    }
+                }
             }
         }
     }
