@@ -3,22 +3,23 @@ package com.example.foregroundservice
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
 import androidx.core.content.ContextCompat
-import kotlinx.android.synthetic.main.activity_main.*
+import com.example.foregroundservice.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
+    val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(binding.root)
 
-        buttonStart.setOnClickListener {
+        binding.buttonStart.setOnClickListener {
             val intent = Intent(this, Foreground::class.java)
             ContextCompat.startForegroundService(this, intent)
         }
 
-        buttonStop.setOnClickListener {
+        binding.buttonStop.setOnClickListener {
             val intent = Intent(this, Foreground::class.java)
             stopService(intent)
         }
