@@ -2,23 +2,27 @@ package com.example.networkhttpurlconnection
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_main.*
+import com.example.networkhttpurlconnection.databinding.ActivityMainBinding
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import java.io.BufferedReader
 import java.io.InputStreamReader
 import java.net.HttpURLConnection
 import java.net.URL
-import kotlin.concurrent.thread
 
 class MainActivity : AppCompatActivity() {
 
+    val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(binding.root)
 
-        buttonRequest.setOnClickListener {
-            thread(start=true) {
+        binding.buttonRequest.setOnClickListener {
+            CoroutineScope(Dispatchers.IO).launch {
                 try {
-                    var urlText = editUrl.text.toString()
+                    var urlText = binding.editUrl.text.toString()
                     if (!urlText.startsWith("https")) {
                         urlText = "https://${urlText}"
                     }
@@ -36,8 +40,8 @@ class MainActivity : AppCompatActivity() {
                         }
                         buffered.close()
                         urlConnection.disconnect()
-                        runOnUiThread {
-                            textContent.text = content.toString()
+                        launch(Dispatchers.Main) {
+                            binding.textContent.text = content.toString()
                         }
                     }
                 } catch (e: Exception) {
