@@ -3,7 +3,7 @@ package com.example.networkretrofit
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
-import kotlinx.android.synthetic.main.activity_main.*
+import com.example.networkretrofit.databinding.ActivityMainBinding
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -11,33 +11,34 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.http.GET
 
+class MainActivity: AppCompatActivity() {
 
-class MainActivity : AppCompatActivity() {
+    val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(binding.root)
 
         val adapter = CustomAdapter()
-        recyclerView.adapter = adapter
-        recyclerView.layoutManager = LinearLayoutManager(this)
+        binding.recyclerView.adapter = adapter
+        binding.recyclerView.layoutManager = LinearLayoutManager(this)
 
         val retrofit = Retrofit.Builder()
-            .baseUrl("https://api.github.com/")
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
+                .baseUrl("https://api.github.com/")
+                .addConverterFactory(GsonConverterFactory.create())
+                .build()
 
-        buttonRequest.setOnClickListener {
+        binding.buttonRequest.setOnClickListener {
             val githubService = retrofit.create(GithubService::class.java)
-            githubService.users().enqueue(object: Callback<List<Repository>> {
-                override fun onFailure(call: Call<List<Repository>>, t: Throwable) {
+            githubService.users().enqueue(object: Callback<Repository> {
+                override fun onFailure(call: Call<Repository>, t: Throwable) {
                     /* */
                 }
                 override fun onResponse(
-                    call: Call<List<Repository>>,
-                    response: Response<List<Repository>>
+                        call: Call<Repository>,
+                        response: Response<Repository>
                 ) {
-                    adapter.userList.addAll(response.body() as List<Repository>)
+                    adapter.userList = response.body() as Repository
                     adapter.notifyDataSetChanged()
                 }
             })
@@ -47,5 +48,5 @@ class MainActivity : AppCompatActivity() {
 
 interface GithubService {
     @GET("users/Kotlin/repos")
-    fun users() : Call<List<Repository>>
+    fun users(): Call<Repository>
 }
