@@ -3,32 +3,34 @@ package com.example.sqlite
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
-import kotlinx.android.synthetic.main.activity_main.*
+import com.example.sqlite.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
+
+    val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
 
     val helper = SqliteHelper(this, "memo", 1)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(binding.root)
 
         val adapter = RecyclerAdapter()
         adapter.helper = helper
 
         adapter.listData.addAll(helper.selectMemo())
-        recyclerMemo.adapter = adapter
-        recyclerMemo.layoutManager = LinearLayoutManager(this)
+        binding.recyclerMemo.adapter = adapter
+        binding.recyclerMemo.layoutManager = LinearLayoutManager(this)
 
-        buttonSave.setOnClickListener {
-            if (editMemo.text.toString().isNotEmpty()) {
-                val memo = Memo(null, editMemo.text.toString(), System.currentTimeMillis())
+        binding.buttonSave.setOnClickListener {
+            if (binding.editMemo.text.toString().isNotEmpty()) {
+                val memo = Memo(null, binding.editMemo.text.toString(), System.currentTimeMillis())
                 helper.insertMemo(memo)
 
                 adapter.listData.clear()
                 adapter.listData.addAll(helper.selectMemo())
                 adapter.notifyDataSetChanged()
-                editMemo.setText("")
+                binding.editMemo.setText("")
             }
         }
     }
