@@ -3,21 +3,23 @@ package net.flow9.viewpager
 import android.os.Bundle
 import androidx.fragment.app.FragmentActivity
 import com.google.android.material.tabs.TabLayoutMediator
-import kotlinx.android.synthetic.main.activity_main.*
+import net.flow9.viewpager.databinding.ActivityMainBinding
 
-// ViewPager 2
 class MainActivity : FragmentActivity() {
+
+    val binding by lazy {ActivityMainBinding.inflate(layoutInflater)}
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(binding.root)
 
         val fragmentList = listOf(FragmentA(), FragmentB(), FragmentC(), FragmentD())
         val adapter = FragmentAdapter(this)
         adapter.fragmentList = fragmentList
-        viewPager.adapter = adapter
+        binding.viewPager.adapter = adapter
 
         val tabTitles = listOf<String>("A", "B", "C", "D")
-        TabLayoutMediator(tabLayout, viewPager) { tab, position ->
+        TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
             tab.text = tabTitles[position]
         }.attach()
     }
