@@ -2,18 +2,33 @@ package com.example.fragment
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.fragment.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
+    val binding by lazy {ActivityMainBinding.inflate(layoutInflater) }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(binding.root)
 
         setFragment()
+
+        binding.btnSend.setOnClickListener {
+            listFragment.setValue("전달할 값")
+        }
     }
 
+    lateinit var listFragment:ListFragment
+
     fun setFragment() {
-        val listFragment:ListFragment = ListFragment()
+        listFragment = ListFragment()
+
+        var bundle = Bundle()
+        bundle.putString("key1", "List Fragment")
+        bundle.putInt("key2", 20170727)
+        listFragment.arguments = bundle
+
         val transaction = supportFragmentManager.beginTransaction()
         transaction.add(R.id.frameLayout, listFragment)
         transaction.commit()
